rename the test file
